fixed all linter errors
